requirements.txt
# Core
anthropic>=0.40.0
httpx>=0.27.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
praw>=7.7.0

# Database
sqlite-utils>=3.35

# CLI & Display
typer>=0.12.0
rich>=13.7.0

# Data & Config
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pandas>=2.2.0

# Utilities
tenacity>=8.2.0
fake-useragent>=1.4.0
